bare specifier in lume task and use jsdelivr
Deno can run now bare specifiers and it's automatically resolved using
the import map (previously this wasn't possible, hence the echo thing).
https://lume.land/blog/posts/lume-3-1-0/#lume-task-uses-bare-specifier

This makes possible the use of permissions
(in order to make the build more secure) and other features related
with stdin in future Lume versions.

jsdelivr is the recommended way. Lume was moved to JsDelivr
https://lume.land/blog/posts/lume-3-1-0/#lume-is-moving-to-jsdelivr

and this will make easier the creation of themes because you can use
glob patterns instead of configuring all files manually
https://lume.land/blog/posts/lume-3-1-0/#create-themes-is-now-easier